visualize_yolo/requirements.txt
flask
torch
numpy
pillow
django
django-admin
torchvision
matplotlib
opencv-python
pandas
requests